cryptov3/backend/requirements.txt
tensorflow>=2.10.0
keras>=2.10.0
opencv-python>=4.5.0
pillow>=8.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
numpy>=1.19.0
pydantic>=1.8.0
pandas>=1.3.0
ta>=0.10.0
scikit-learn>=1.0.0
xgboost>=1.5.0
statsmodels>=0.13.0
matplotlib>=3.5.0
